+++ tests/psa-client-server/psasim/include/service.h
@@ -15,7 +15,7 @@ extern "C" {
 #include <stdint.h>
 #include <stddef.h>
 
-#include "psa/common.h"
+#include "common.h"
 
 #include "psa/crypto.h"
